+++ ASCOfficeRtfFile/RtfFormatLib/source/Reader/OOXDocumentReader.h
@@ -71,7 +71,7 @@ public:
 		int last_section_start = 0;
 		
 		//считаем количесво секций и заполняем их свойства .. 
-		for (long i = 0; i < m_ooxDocument->m_arrItems.size(); i++)
+		for (size_t i = 0; i < m_ooxDocument->m_arrItems.size(); i++)
 		{
 			if (m_ooxDocument->m_arrItems[i] == NULL) continue;
 
@@ -86,7 +86,7 @@ public:
                         RtfSectionPtr s(new RtfSection());
                         _section section(s, last_section_start, i + 1);
 
-						last_section_start = i + 1;
+						last_section_start = (int)i + 1;
 
 						section.props->m_oProperty.SetDefaultOOX();
 
@@ -106,7 +106,7 @@ public:
 		if (last_section.end_para < m_ooxDocument->m_arrItems.size())
 		{
             RtfSectionPtr s(new RtfSection());
-            _section section(s, last_section.end_para, m_ooxDocument->m_arrItems.size());
+            _section section(s, last_section.end_para, (int)m_ooxDocument->m_arrItems.size());
 			
 			section.props->m_oProperty.SetDefaultOOX();
 			if (m_ooxDocument->m_oSectPr.IsInit())// свойства последней секции
